frontend/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
requests==2.31.0
numpy==1.26.4
Pillow==10.3.0
prometheus_client==0.21.0
